backend/tests/api/test_api.py
import pytest
from fastapi.testclient import TestClient
from unittest.mock import patch, Mock

from app.main import app


@pytest.fixture
def client():
    """Create a test client."""
    return TestClient(app)


class TestMainApp:
    """Test cases for the main FastAPI application."""
    
    def test_root_endpoint(self, client):
        """Test the root endpoint."""
        response = client.get("/")
        assert response.status_code == 200
        assert response.json() == {"message": "Security Vulnerabilities Knowledge Base API"}
    
    def test_health_check_endpoint(self, client):
        """Test the health check endpoint."""
        response = client.get("/api/health")
        assert response.status_code == 200
        data = response.json()
        assert data["status"] == "healthy"
    
    def test_cors_headers(self, client):
        """Test that CORS headers are properly set."""
        response = client.options("/api/health")
        # The test client doesn't automatically handle preflight requests
        # but we can test that the app is configured with CORS middleware
        assert response.status_code in [200, 405]  # 405 is OK for OPTIONS on GET endpoints
    
    @patch('app.main.run_full_job')
    def test_startup_event_triggers_scraping(self, mock_run_full_job, client):
        """Test that the startup event triggers initial scraping."""
        # The startup event should have been triggered when creating the test client
        # However, testing startup events with TestClient can be tricky
        # This test verifies the endpoint structure is correct
        response = client.get("/")
        assert response.status_code == 200
    
    def test_api_prefix_configuration(self, client):
        """Test that API endpoints are correctly prefixed."""
        # Test that endpoints are available under /api prefix
        response = client.get("/api/health")
        assert response.status_code == 200
        
        # Test that direct access without prefix fails
        response = client.get("/health")
        assert response.status_code == 404
    
    def test_invalid_endpoint(self, client):
        """Test accessing invalid endpoints."""
        response = client.get("/nonexistent")
        assert response.status_code == 404
        
        response = client.get("/api/nonexistent")
        assert response.status_code == 404