backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.1
chromadb==0.4.22
openai==1.14.0
numpy==1.26.3
pydantic==2.6.1
langchain>=0.1.0,<0.2.0
langchain-text-splitters>=0.0.1
python-multipart==0.0.6
schedule==1.2.1
httpx==0.27.2